requirements.txt
mock==2.0.0
pbr==1.9.1
py==1.4.31
PyJWT==1.4.0
pytest==2.9.1
requests==2.10.0
requests-mock==1.0.0
six==1.10.0
